+++ FJ-VDMJ4/src/main/java/com/fujitsu/vdmj/in/statements/INMapSeqDesignator.java
@@ -88,7 +88,7 @@ public class INMapSeqDesignator extends INStateDesignator
 			else if (root.isType(SeqValue.class))
 			{
 				ValueList seq = root.seqValue(ctxt);
-				int i = (int)index.intValue(ctxt)-1;
+				int i = (int)index.intValue(ctxt).intValue()-1;
 
 				if (!seq.inbounds(i))
 				{
