+++ FJ-VDMJ4/src/main/java/com/fujitsu/vdmj/values/UpdatableValue.java
@@ -23,6 +23,9 @@
 
 package com.fujitsu.vdmj.values;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import com.fujitsu.vdmj.Settings;
 import com.fujitsu.vdmj.config.Properties;
 import com.fujitsu.vdmj.lex.Dialect;
@@ -226,25 +229,25 @@ public class UpdatableValue extends ReferenceValue
 	}
 
 	@Override
-	public synchronized double realValue(Context ctxt) throws ValueException
+	public synchronized BigDecimal realValue(Context ctxt) throws ValueException
 	{
 		return value.realValue(ctxt);
 	}
 
 	@Override
-	public synchronized long intValue(Context ctxt) throws ValueException
+	public synchronized BigInteger intValue(Context ctxt) throws ValueException
 	{
 		return value.intValue(ctxt);
 	}
 
 	@Override
-	public synchronized long natValue(Context ctxt) throws ValueException
+	public synchronized BigInteger natValue(Context ctxt) throws ValueException
 	{
 		return value.nat1Value(ctxt);
 	}
 
 	@Override
-	public synchronized long nat1Value(Context ctxt) throws ValueException
+	public synchronized BigInteger nat1Value(Context ctxt) throws ValueException
 	{
 		return value.nat1Value(ctxt);
 	}
