+++ FJ-VDMJUnit4/src/main/java/com/fujitsu/vdmjunit/VDMJUnitTest.java
@@ -25,6 +25,8 @@ package com.fujitsu.vdmjunit;
 
 import static org.junit.Assert.fail;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.nio.charset.Charset;
 
 import com.fujitsu.vdmj.Release;
@@ -213,7 +215,7 @@ public abstract class VDMJUnitTest
 	 * 			or the specification has not been initialized, or if the result is not
 	 * 			a VDM integer value.
 	 */
-	protected long runInt(String expression) throws Exception
+	protected BigInteger runInt(String expression) throws Exception
 	{
 		return run(expression).intValue(null);
 	}
@@ -228,7 +230,7 @@ public abstract class VDMJUnitTest
 	 * 			or the specification has not been initialized, or if the result is not
 	 * 			a VDM real value.
 	 */
-	protected double runReal(String expression) throws Exception
+	protected BigDecimal runReal(String expression) throws Exception
 	{
 		return run(expression).realValue(null);
 	}
